migrate response handling to trait. update ACL roles
